cms-admin: Move ContentScope stories to package Storybook
Move stories for ContentScopeProvider, ContentScopeSelect and NoContentScopeFallback
from the global /storybook into the dedicated Storybook of @comet/cms-admin.
